+++ install.sh
@@ -1,17 +1,5 @@
 #!/usr/bin/env bash
 
-# constants
-# DEFAULT_PYVER=3.12
-
-usage()
-{
-#    echo "Usage: install.sh  [ -p Python version (3.12) ]
-    echo "Usage: install.sh  
-                  [ -d  Install developer tools ]
-            "
-    exit 2
-}
-
 unamestr=`uname`
 if [ "$unamestr" == 'Linux' ]; then
     prof=~/.bashrc
@@ -24,15 +12,11 @@ else
     exit
 fi
 
-
 # execute the user's profile
 source $prof
 
 
-# Parse the command line arguments passed in by the user
-# PYVER=$DEFAULT_PYVER
-
-# Do mac-specific conda installs
+# Do os-specific conda dependencies
 if [ "$unamestr" == 'FreeBSD' ] || [ "$unamestr" == 'Darwin' ]; then
     # This is motivated by the mysterios pyproj/rasterio error and incorrect results
     # that only happen on ARM macs. 
@@ -42,30 +26,40 @@ else
     input_yaml_file=linux_environment.yml
 fi
 
-developer=false
+
+# Set some default arguments
 # Default is to use conda to install since mamba fails on some systems
 install_pgm=conda
-while getopts "p:d" options; do
-    case "${options}" in 
-    d)
+developer=false
+VENV=shakemap
+
+# Arguments
+# usage() { echo "Usage: $0 [-d to install developer dependencies] [-n <environment name>]" 1>&2; exit 1; }
+usage() { echo "$0 usage:" && grep " .)\ #" $0; exit 0; }
+
+while getopts ":hdn:" options; do
+    case $options in 
+    d) # Install additional developer dependencies.
         developer=true
         ;;
-    *)                            # If unknown (any other) option:
-      usage                       # Exit abnormally.
+    n) # Overwrite name of virtual environment.
+        VENV="$OPTARG"
+        ;;
+    h | *) # Display help.
+      usage
+      exit 0
       ;;
     esac
 done
 
 echo "YAML file to use as input: ${input_yaml_file}"
-#echo "Using python version ${PYVER}"
 
-# Name of virtual environment, pull from yml file
-VENV=`grep "name:" source_environment.yml  | cut -f2 -d ":" | sed 's/ //g'`
-echo "#####Environment to create: '${VENV}'"
+# Name of virtual environment
+echo "Environment to create: '${VENV}'"
 
 # Where is conda installed?
 CONDA_LOC=`which conda`
-echo "######Location of conda install: ${CONDA_LOC}"
+echo "Location of conda install: ${CONDA_LOC}"
 
 # Are we in an environment
 CURRENT_ENV=`conda info --envs | grep "*"`
@@ -92,39 +86,29 @@ elif [ ! grep -Fxq "backend" $matplotlibrc ]; then
     echo "NOTE: A non-interactive matplotlib backend (Agg) has been set for this user."
 else
     sed -i '' 's/backend.*/backend : Agg/' $matplotlibrc
-    echo "###############"
     echo "NOTE: $matplotlibrc has been changed to set 'backend : Agg'"
-    echo "###############"
 fi
 
 # Is conda installed?
 conda --version
 if [ $? -ne 0 ]; then
     echo "No conda detected, installing miniconda..."
-
     command -v curl >/dev/null 2>&1 || { echo >&2 "Script requires curl but it's not installed. Aborting."; exit 1; }
-
     miniforge_url="https://github.com/conda-forge/miniforge/releases/latest/download/Miniforge3-$(uname)-$(uname -m).sh"
     curl -L $miniforge_url -o miniforge.sh &>/dev/null
-
     # if curl fails, bow out gracefully
     if [ $? -ne 0 ];then
         echo "Failed to download miniconda installer shell script. Exiting."
         exit 1
     fi
-    
     echo "Install directory: $HOME/miniconda"
-
     bash miniforge.sh -f -b -p $HOME/miniconda &>/dev/null
-
     # if miniforge.sh fails, bow out gracefully
     if [ $? -ne 0 ];then
         echo "Failed to run miniconda installer shell script. Exiting."
         exit 1
     fi
-    
     . $HOME/miniconda/etc/profile.d/conda.sh
-
     # remove the shell script
     rm miniforge.sh
 else
@@ -138,13 +122,13 @@ LATEST=`conda search conda | tail -1 | tr -s ' ' | cut -f2 -d" "`
 echo "${CVNUM}"
 echo "${LATEST}"
 if [ ${LATEST} != ${CVNUM} ]; then
-    echo "##################Updating conda tool..."
+    echo "Updating conda tool..."
     CVERSION=`conda -V`
     echo "Current conda version: ${CVERSION}"
     conda update -n base conda -y
     CVERSION=`conda -V`
     echo "New conda version: ${CVERSION}"
-    echo "##################Done updating conda tool..."
+    echo "Done updating conda tool..."
 else
     echo "conda ${CVNUM} already matches latest version ${LATEST}. No update required."
 fi
@@ -160,24 +144,19 @@ echo "Activate base virtual environment"
 # whatever it does, it is crucially important for being able to activate a conda environment
 # inside a shell script.
 eval "$(conda shell.bash hook)"                                                
-conda activate base
+${install_pgm} activate base
 if [ $? -ne 0 ]; then
     "Failed to activate conda base environment. Exiting."
     exit 1
 fi
 
 # Remove existing shakemap environment if it exists
-conda remove -y -n $VENV --all
-conda clean -y --all
+${install_pgm} remove -y -n $VENV --all
+${install_pgm} clean -y --all
 
 # Install the virtual environment
-# echo "Creating new environment from environment file: ${input_yaml_file} with python version ${PYVER}"
 echo "Creating new environment from environment file: ${input_yaml_file}"
-# change python version in yaml file to match PYVER
-# sed 's/python='"${DEFAULT_PYVER}"'/python='"${PYVER}"'/' "${input_yaml_file}" > tmp.yml
-# ${install_pgm} env create -f tmp.yml
-${install_pgm} env create -f ${input_yaml_file}
-# rm tmp.yml 
+${install_pgm} env create -f ${input_yaml_file} -n $VENV
 
 
 # Bail out at this point if the conda create command fails.
@@ -189,7 +168,7 @@ fi
 
 # Activate the new environment
 echo "Activating the $VENV virtual environment"
-conda activate $VENV
+${install_pgm} activate $VENV
 
 # if conda activate fails, bow out gracefully
 if [ $? -ne 0 ];then
@@ -197,27 +176,22 @@ if [ $? -ne 0 ];then
     exit 1
 fi
 
-# The presence of a __pycache__ folder in bin/ can cause the pip
-# install to fail... just to be safe, we'll delete it here.
-if [ -d bin/__pycache__ ]; then
-    rm -rf bin/__pycache__
-fi
-
-
 
 if $developer; then
-    echo "############# Installing shakemap with developer tools ##############"
+    echo "Installing shakemap with developer tools."
     conda install mathjax -y
     if ! pip install -e '.[dev,test,doc]' ; then
         echo "Installation of shakemap failed."
         exit 1
     fi
 else
-    echo "############# Installing shakemap ##############"
+    echo "Installing shakemap."
     if ! pip install -e . ; then
         echo "Installation of shakemap failed."
         exit 1
     fi
 fi
 
-echo "Reminder: Run 'conda activate shakemap' to enable the ShakeMap environment."
+echo "*********"
+echo "Reminder: Run 'conda activate ${VENV}' to enable the environment."
+echo "*********"